stay with simple furigana for expressions only as of now
creating automated furigana for entire sentences with no base is difficult since there are words with several readings and meanings based on those.  A script can't detect whether ひらく or　あく is needed for 開く so for all used kanji in example sentences I would need the reading. Which I have, but it isn't matching Sudachi's tokenizing of the sentence string.